refactor: move get_socket_path to utils
