+++ ulauncher/utils/socket_path.py
